# Conflicts:
#	src/Parse.java

+++ src/ReadFile.java
@@ -3,6 +3,7 @@ import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
+import sun.awt.Mutex;
 
 
 import java.io.*;
@@ -10,6 +11,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
 
 import static java.nio.file.StandardOpenOption.APPEND;
 import static java.nio.file.StandardOpenOption.CREATE;
@@ -17,28 +19,30 @@ import static java.nio.file.StandardOpenOption.CREATE;
 public class ReadFile implements Runnable{
 
     private Path pathData;
-    private volatile Integer fileNumber ;
-    private Parse parser;
+
     private File folder ;
-    private volatile HashSet<String> filesExecuted ;
-    private volatile Object lookIncrease;
-    private Map<String , Map<Integer,Integer>> Voc ;
-    private Map<Integer , String> fileMeta_data;
-    private int fileIteration ;
+    private static volatile ConcurrentHashMap<String,Integer> filesExecuted = new ConcurrentHashMap<>();
+    private static volatile Object lookClean  = new Object();
+    private static volatile boolean cleanfile = true;
+    private Indexer ind ;
+
 
     public ReadFile(Path pathData , final File folder) {
-        if (Files.exists(pathData)) {
-            // file exist
-            System.out.println("exist");
-            deleteDirectory(new File(pathData.toString()));
-        }
+            synchronized (lookClean) {
+                if (cleanfile) {
+                    cleanfile = false;
+                    if (Files.exists(pathData)) {
+                        // file exist
+                        System.out.println("exist");
+                        deleteDirectory(new File(pathData.toString()));
+                    }
 
+                }
+
+            }
         new File(pathData.toString()).mkdirs();
         this.pathData = pathData;
-        fileNumber=0;
-        this.folder=folder;
-        this.filesExecuted = new HashSet<>();
-        lookIncrease = new Object();
+        this.folder = folder;
     }
 
     private boolean deleteDirectory(File directoryToBeDeleted) {
@@ -57,9 +61,9 @@ public class ReadFile implements Runnable{
                 listFilesForFolder(fileEntry);
             } else {
                 boolean check = false ;
-                synchronized (this) {
-                    if (!filesExecuted.contains(fileEntry.toString())) {
-                        filesExecuted.add(fileEntry.toString());
+                synchronized (lookClean) {
+                    if (!filesExecuted.containsKey(fileEntry.toPath().toString())) {
+                        filesExecuted.put(fileEntry.toPath().toString(),0);
                         check = true ;
                     }
                 }
@@ -72,47 +76,12 @@ public class ReadFile implements Runnable{
     }
 
     public void readFile(Path file ,String fileName) {
-//        try (InputStream in = Files.newInputStream(file);
-//             BufferedReader reader =
-//                     new BufferedReader(new InputStreamReader(in))) {
-//            String line = null;
-//            String text="" ;
-        Voc = new HashMap<>();
+
+
         try {
                   Document document = Jsoup.parse(new String(Files.readAllBytes(file)));
                   Elements All_docs =  document.getElementsByTag("DOC");
-            fileMeta_data = new HashMap<>();
-            for (Element doc:
-                 All_docs) {
-                Integer local_file_num;
-                synchronized (lookIncrease) {
-                    local_file_num = fileNumber;
-                    fileNumber++;
-                }
-                Elements title =  doc.getElementsByTag("TI");
-                Elements date =  doc.getElementsByTag("DATE1");
-                Elements docno =  doc.getElementsByTag("DOCNO");
-                fileMeta_data.put(local_file_num , "="+title.text()+"="+date.text()+"="+docno.text());
-
-                Map<String,Integer> mp = parser.parseIt(doc.text());
-                Iterator<String> it = mp.keySet().iterator();
-                while (it.hasNext()){
-                    String s = it.next();
-                    if(Voc.get(s)==null){
-                        Map<Integer,Integer>  l= new HashMap<>();
-                        Voc.put(s , l);
-                    }
-                    Voc.get(s).put(local_file_num , mp.get(s));
-                }
-                fileIteration++;
-                if(fileIteration>1000){
-                    fileIteration = 0 ;
-                    writeFile(fileMeta_data.toString() , local_file_num.toString());
-                    writeFile(Voc.toString() , local_file_num.toString()+"Voc");
-                    fileMeta_data = new HashMap<>();
-                    Voc = new HashMap<>();
-                }
-            }
+                  ind.WriteData(All_docs , false);
 
         }catch (IOException x) {
             System.err.println(x);
@@ -132,26 +101,12 @@ public class ReadFile implements Runnable{
 //        }
     }
 
-    public void writeFile(String s , String filename){
-        byte data[] = s.getBytes();
-        String st ;
-        synchronized (lookIncrease) {
-         //   new File(pathData.toString() + "//@" + fileNumber).mkdirs();
-            st = this.pathData.toString() + "//@"  + filename;
-        }
-        System.out.println(st);
-        Path p = Paths.get(st);
-        try (OutputStream out = new BufferedOutputStream(
-                Files.newOutputStream(p, CREATE))) {
-            out.write(data, 0, data.length);
-        } catch (IOException x) {
-            System.err.println(x);
-        }
-    }
+
 
     @Override
     public void run() {
-        parser = new Parse();
+        this.ind = new Indexer(pathData);
         listFilesForFolder(folder);
+        ind.WriteData(null , true);
     }
 }
